pyproject.toml
[project]
name = "video-editor"
version = "0.1.0"
description = "Video editing scripts for Marky"
requires-python = ">=3.10"
dependencies = [
    "deepgram-sdk>=5.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "prefect>=3.0.0",
    "google-genai>=0.2.0",
    "pillow>=10.0.0",
    "streamlit>=1.30.0",
    "opencv-python-headless>=4.8.0",
    "mediapipe>=0.10.0",
    "openai>=1.50.0",
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",
    "rich>=13.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiosqlite>=0.20.0",
]

[tool.uv]
package = false
